Remove redundant NULL-after-free, rename const_ files, fix stale comments
Post-refactor cleanup after the Constant→Parameter merge:
- Remove dead NULL assignments in free_type_data (quad_form, hstack, index, linear_op, left_matmul, scalar_mult, vector_mult)
- Rename const_scalar_mult/const_vector_mult source and test files to drop const_ prefix
- Rename test_variable_constant.h → test_variable_parameter.h and test_constant → test_fixed_parameter
- Update stale comments in multiply.c and bivariate.h
- Rename const_scalar_mult_expr/const_vector_mult_expr structs to scalar_mult_expr/vector_mult_expr
- Update left_matmul param_source comment
