+++ src/affine/hstack.c
@@ -165,13 +165,10 @@ static void free_type_data(expr *node)
     for (int i = 0; i < hnode->n_args; i++)
     {
         free_expr(hnode->args[i]);
-        hnode->args[i] = NULL;
     }
 
     free_csr_matrix(hnode->CSR_work);
-    hnode->CSR_work = NULL;
     free(hnode->args);
-    hnode->args = NULL;
 }
 
 expr *new_hstack(expr **args, int n_args, int n_vars)
