+++ tests/forward_pass/affine/test_variable_parameter.h
@@ -18,13 +18,13 @@ const char *test_variable()
     return 0;
 }
 
-const char *test_constant()
+const char *test_fixed_parameter()
 {
     double c[2] = {5.0, 10.0};
     double u[2] = {0.0, 0.0};
     expr *const_node = new_parameter(2, 1, PARAM_FIXED, 0, c);
     const_node->forward(const_node, u);
-    mu_assert("Constant test failed", cmp_double_array(const_node->value, c, 2));
+    mu_assert("Fixed parameter test failed", cmp_double_array(const_node->value, c, 2));
     free_expr(const_node);
     return 0;
 }
